+++ src/main/java/domain/actions/GoingToChargingStationAction.java
@@ -34,13 +34,13 @@ public class GoingToChargingStationAction extends GraphDefinedDomain.GraphAction
 
 
     @Override
-    public double getActionTime(TaxiGraphState state) {
+    public int getActionTime(TaxiGraphState state) {
         return getTripTime(state.getNodeId(), toNodeId) ;
     }
 
 
     @Override
-    public double getActionEnergyConsumption(TaxiGraphState state) {
+    public int getActionEnergyConsumption(TaxiGraphState state) {
         return ActionUtils.getActionEnergyConsumption(state, toNodeId, getActionTime(state));
     }
